requirements.txt
Django==5.1.4
djangorestframework==3.15.2
python-dotenv==1.0.1
django-cors-headers==4.6.0
djangorestframework-simplejwt==5.4.0
drf_yasg==1.21.8